phase4/requirements.txt
pystray==0.19.5
Pillow==10.3.0
gspread==6.1.2
google-auth==2.29.0
